Classes that inherit sealed class should not be in a one file anymore
### What's done:
- separated tree node classes
