+++ ktoml-core/src/commonMain/kotlin/com/akuleshov7/ktoml/encoders/TomlMainEncoder.kt
@@ -2,7 +2,10 @@ package com.akuleshov7.ktoml.encoders
 
 import com.akuleshov7.ktoml.TomlInputConfig
 import com.akuleshov7.ktoml.TomlOutputConfig
-import com.akuleshov7.ktoml.tree.*
+import com.akuleshov7.ktoml.tree.nodes.pairs.keys.TomlKey
+import com.akuleshov7.ktoml.tree.nodes.*
+import com.akuleshov7.ktoml.tree.nodes.pairs.values.TomlArray
+import com.akuleshov7.ktoml.tree.nodes.pairs.values.TomlValue
 import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.SerializationStrategy
 import kotlinx.serialization.descriptors.SerialDescriptor
