+++ ktoml-core/src/commonMain/kotlin/com/akuleshov7/ktoml/parsers/TomlParser.kt
@@ -3,6 +3,7 @@ package com.akuleshov7.ktoml.parsers
 import com.akuleshov7.ktoml.TomlConfig
 import com.akuleshov7.ktoml.TomlInputConfig
 import com.akuleshov7.ktoml.tree.*
+import com.akuleshov7.ktoml.tree.nodes.*
 import kotlin.jvm.JvmInline
 
 /**
