+++ ktoml-core/src/commonMain/kotlin/com/akuleshov7/ktoml/tree/nodes/TomlFile.kt
@@ -1,4 +1,4 @@
-package com.akuleshov7.ktoml.tree
+package com.akuleshov7.ktoml.tree.nodes
 
 import com.akuleshov7.ktoml.TomlConfig
 import com.akuleshov7.ktoml.TomlInputConfig
@@ -24,16 +24,16 @@ public class TomlFile(config: TomlInputConfig = TomlInputConfig()) : TomlNode(
     public constructor(config: TomlConfig) : this(config.input)
 
     override fun getNeighbourNodes(): MutableList<TomlNode> =
-            throw InternalAstException("Invalid call to getNeighbourNodes() for TomlFile node")
+        throw InternalAstException("Invalid call to getNeighbourNodes() for TomlFile node")
 
     override fun write(
         emitter: TomlEmitter,
         config: TomlOutputConfig,
         multiline: Boolean
     ): Unit =
-            emitter.writeChildren(
-                children,
-                config,
-                multiline
-            )
+        emitter.writeChildren(
+            children,
+            config,
+            multiline
+        )
 }
