+++ ktoml-core/src/commonMain/kotlin/com/akuleshov7/ktoml/tree/nodes/pairs/values/TomlDouble.kt
@@ -1,0 +1,26 @@
+package com.akuleshov7.ktoml.tree.nodes.pairs.values
+
+import com.akuleshov7.ktoml.TomlOutputConfig
+import com.akuleshov7.ktoml.writers.TomlEmitter
+
+/**
+ * Toml AST Node for a representation of float types: key = 1.01.
+ * Toml specification requires floating point numbers to be IEEE 754 binary64 values,
+ * so it should be Kotlin Double (64 bits)
+ * @property content
+ */
+public class TomlDouble
+internal constructor(
+    override var content: Any,
+    lineNo: Int
+) : TomlValue(lineNo) {
+    public constructor(content: String, lineNo: Int) : this(content.toDouble(), lineNo)
+
+    override fun write(
+        emitter: TomlEmitter,
+        config: TomlOutputConfig,
+        multiline: Boolean
+    ) {
+        emitter.emitValue(content as Double)
+    }
+}
